Removed debugging utilities.
