--- example_apps/transit/transit-fixture/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "transit-fixture"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tokio = { workspace = true, features = ["time", "macros", "sync"]}
-serde = { workspace = true, features = ["derive"]}
-quick-xml = { version = "0.30", features = ["serialize", "serde-types"] }
-thiserror = { workspace = true }
-hyper = { workspace = true }
-tracing = { workspace = true }
-transit-model = { path = "../transit-model" }
-rand = { workspace = true }
-axum = { version = "0.6.20", features = ["tokio", "query", "http1"]}
-bytes = { workspace = true }
-parking_lot = { workspace = true }
-chrono = { workspace = true }
-futures = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["time", "macros", "sync", "signal", "rt-multi-thread"]}
